Add similarity percentage to debug output
This commit enhances the debug output of the PHP file comparison tool by adding a similarity percentage.

The similarity is calculated using the `similar_text` function and is displayed when the files are not equal and debugging is enabled. This provides a quick indication of how similar the files are, despite their differences.